# ApplicationTracker.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ApplicationTracker.Api",
    "version": "v1"
  },
  "paths": {
    "/Application/new": {
      "post": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NewJobApplicationDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NewJobApplicationDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NewJobApplicationDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/new/multiple": {
      "post": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewJobApplicationDto"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewJobApplicationDto"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewJobApplicationDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/userApplications": {
      "get": {
        "tags": [
          "Application"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Application/application": {
      "get": {
        "tags": [
          "Application"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              }
            }
          }
        }
      }
    },
    "/Application/delete": {
      "delete": {
        "tags": [
          "Application"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/delete/multiple": {
      "delete": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/update": {
      "patch": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JobApplicationDisplayDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/update/multiple": {
      "patch": {
        "tags": [
          "Application"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobApplicationDisplayDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Application/temp": {
      "get": {
        "tags": [
          "Application"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/jwt": {
      "post": {
        "tags": [
          "Debug tool to check token content"
        ],
        "parameters": [
          {
            "name": "jwt",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "JobApplicationDisplayDto": {
        "required": [
          "company",
          "jobApplicationId",
          "jobTitle"
        ],
        "type": "object",
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "uuid"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "applicationPdf": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PdfFile"
              }
            ],
            "nullable": true
          },
          "resumePdf": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PdfFile"
              }
            ],
            "nullable": true
          },
          "appliedDate": {
            "type": "string",
            "format": "date-time"
          },
          "replyDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "NewJobApplicationDto": {
        "required": [
          "company",
          "jobTitle"
        ],
        "type": "object",
        "properties": {
          "jobTitle": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "company": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "applicationPdf": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PdfFileDto"
              }
            ],
            "nullable": true
          },
          "resumePdf": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PdfFileDto"
              }
            ],
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PdfFile": {
        "required": [
          "pdfType"
        ],
        "type": "object",
        "properties": {
          "pdfFileId": {
            "type": "string",
            "format": "uuid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "uuid"
          },
          "pdfType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PdfType"
              }
            ]
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PdfFileDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PdfType": {
        "enum": [
          "Application",
          "Resume"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { }
  ],
  "tags": [
    {
      "name": "Application"
    },
    {
      "name": "Debug tool to check token content"
    }
  ]
}